Just adding all the code I wrote in the last few days, will clean failed prototypes next.
